+++ src/main/java/com/javaweb/repository/PropertyImageRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PropertyImageRepository extends JpaRepository<PropertyImage, Long> {
+public interface PropertyImageRepository extends JpaRepository<PropertyImage, Integer> {
     List<PropertyImage> findByProperty(PropertyEntity property);
 
     interface ListingRepository {
